+++ msat/include/msat_sort.h
@@ -16,7 +16,7 @@ class MsatSort : public AbsSort
   MsatSort(msat_env e, msat_type t, msat_decl d) : env(e), type(t), uf_decl(d), is_uf_type(true) {};
   ~MsatSort() = default;
   std::size_t hash() const override;
-  unsigned int get_width() const override;
+  uint64_t get_width() const override;
   Sort get_indexsort() const override;
   Sort get_elemsort() const override;
   SortVec get_domain_sorts() const override;
